requirements.txt
discord.py==1.7.1
PyNaCl==1.4.0